requirements.txt
discord.py
python-dotenv
imap_tools
beautifulsoup4
mysql-connector-python
